2404update

+++ PMS_API/Models/GoalRating.cs
@@ -8,9 +8,9 @@ namespace PMS_API.Models
         public int RatingId { get; set; }
         public int? EmployeeId { get; set; }
         public int? ManagerId { get; set; }
-        public int? RatingbyEmployee { get; set; }
+        public decimal? RatingbyEmployee { get; set; }
         public DateTime? RatingbyEmployeeCalculatedAt { get; set; }
-        public int? RatingbyManager { get; set; }
+        public decimal? RatingbyManager { get; set; }
         public DateTime? RatingbyManagerCalculatedAt { get; set; }
 
         public virtual EmployeeModule? Employee { get; set; }
